#355 example refactoring of zebra
- for multiple Devices (for each type of scan)
- but hopefully without too much duplication
