+++ src/artemis/devices/system_tests/test_zebra_system.py
@@ -1,25 +1,25 @@
 import pytest
 
-from artemis.devices.zebra import (
+from artemis.devices.zebras.FGSzebra import FGSZebra
+from artemis.devices.zebras.zebra_base import (
     IN3_TTL,
     IN4_TTL,
     OR1,
     PC_PULSE,
     TTL_DETECTOR,
     TTL_SHUTTER,
-    Zebra,
 )
 
 
 @pytest.fixture()
 def zebra():
-    zebra = Zebra(name="zebra", prefix="BL03S-EA-ZEBRA-01:")
+    zebra = FGSZebra(name="zebra", prefix="BL03S-EA-ZEBRA-01:")
     yield zebra
     zebra.pc.disarm().wait(10.0)
 
 
 @pytest.mark.s03
-def test_arm(zebra: Zebra):
+def test_arm(zebra: FGSZebra):
     assert not zebra.pc.is_armed()
     zebra.pc.arm().wait(10.0)
     assert zebra.pc.is_armed()
@@ -27,7 +27,7 @@ def test_arm(zebra: Zebra):
 
 
 @pytest.mark.s03
-def test_disarm(zebra: Zebra):
+def test_disarm(zebra: FGSZebra):
     zebra.pc.arm().wait(10.0)
     assert zebra.pc.is_armed()
     zebra.pc.disarm().wait(10.0)
@@ -35,14 +35,14 @@ def test_disarm(zebra: Zebra):
 
 
 @pytest.mark.s03
-def test_zebra_stage(zebra: Zebra):
+def test_zebra_stage(zebra: FGSZebra):
     zebra.stage()
     assert zebra.output.out_pvs[TTL_DETECTOR].get() == IN3_TTL
     assert zebra.output.out_pvs[TTL_SHUTTER].get() == IN4_TTL
 
 
 @pytest.mark.s03
-def test_zebra_unstage(zebra: Zebra):
+def test_zebra_unstage(zebra: FGSZebra):
     zebra.unstage()
     assert zebra.output.out_pvs[TTL_DETECTOR].get() == PC_PULSE
     assert zebra.output.out_pvs[TTL_SHUTTER].get() == OR1
